+++ crates/ruff_linter/src/checkers/information_flow/information_flow_state.rs
@@ -1,20 +1,16 @@
-use std::{collections::VecDeque, str::FromStr, vec};
+use rustc_hash::FxHashMap;
+use std::collections::VecDeque;
 
-use lazy_static::lazy_static;
-use regex::Regex;
 use ruff_python_index::Indexer;
 use ruff_python_semantic::BindingId;
 use ruff_python_trivia::CommentRanges;
 use ruff_source_file::Locator;
 use ruff_text_size::{TextRange, TextSize};
-use rustc_hash::FxHashMap;
 
-lazy_static! {
-    static ref PRINCIPAL_REGEX: Regex =
-        Regex::new(r"ifprincipals\s*\{\s*(?P<principals>[\w\s,]+)\s*\}").unwrap();
-    static ref LABEL_REGEX: Regex =
-        Regex::new(r"iflabel\s*\{\s*(?P<label>[\w\s,]+)?\s*\}").unwrap();
-}
+use super::{
+    label::Label,
+    principals::{initiate_principals, Principals},
+};
 
 /// State of the information flow
 #[derive()]
@@ -26,11 +22,6 @@ pub(crate) struct InformationFlowState {
     pc: VecDeque<String>,
     // Map from variable name to
     variable_map: FxHashMap<BindingId, Label>,
-}
-
-#[derive(Debug, PartialEq, Clone, Default)]
-pub(crate) struct Label {
-    principals: Vec<String>,
 }
 
 impl InformationFlowState {
@@ -100,229 +91,6 @@ impl InformationFlowState {
                 }
             }
         }
-    }
-}
-
-impl Label {
-    #[allow(dead_code)]
-    pub(crate) fn new(principals: Vec<String>) -> Self {
-        Self { principals }
-    }
-
-    pub(crate) fn is_public(&self) -> bool {
-        self.principals.is_empty()
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn new_public() -> Self {
-        Self { principals: vec![] }
-    }
-}
-
-impl FromStr for Label {
-    fn from_str(string: &str) -> Result<Self, Self::Err> {
-        match LABEL_REGEX.captures(string) {
-            Some(captures) => match captures.name("label") {
-                Some(label) => {
-                    let principals = label
-                        .as_str()
-                        .split(',')
-                        .map(|s| s.trim().to_string())
-                        .filter(|s| !s.is_empty())
-                        .collect::<Vec<String>>();
-                    return Ok(Label { principals });
-                }
-                None => return Ok(Label::new_public()),
-            },
-            None => Err(()),
-        }
-    }
-
-    type Err = ();
-}
-
-/// Check labels direction convertion i.e. you can move down in the lattice, not up
-pub(crate) fn can_convert_label(from_label: &Label, to_label: &Label) -> bool {
-    // If the test label is public, then it is never more restrictive
-    if to_label.is_public() {
-        return true;
-    }
-
-    // If the to_label has more principals, then it is more restrictive
-    if from_label.principals.len() > to_label.principals.len() {
-        // Check if the to_label is a subset of the from_label
-        for principal in &to_label.principals {
-            if !from_label.principals.contains(principal) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    // If the conv_label has the same principals, then it is not more restrictive
-    return to_label.principals.len() == from_label.principals.len()
-        && to_label.principals != from_label.principals;
-}
-
-#[derive(Debug, PartialEq)]
-struct Principals {
-    principals: Vec<String>,
-}
-
-impl Principals {
-    #[allow(dead_code)]
-    fn new(principals: Vec<String>) -> Self {
-        Self { principals }
-    }
-
-    #[allow(dead_code)]
-    fn new_from_str(principals: Vec<&str>) -> Self {
-        Self {
-            principals: principals.iter().map(|s| s.to_string()).collect(),
-        }
-    }
-
-    fn new_empty() -> Self {
-        Self { principals: vec![] }
-    }
-
-    fn concat(&mut self, other: &Principals) {
-        self.principals.extend(other.principals.clone());
-    }
-}
-
-impl FromStr for Principals {
-    /// Parses a string of principals e.g. from a comment
-    /// ```
-    /// ifprincipals {
-    ///   alice,
-    ///   bob
-    /// }
-    /// ```
-    ///
-    /// into the Principals struct with the principals `['alice', 'bob']`
-    fn from_str(string: &str) -> Result<Self, Self::Err> {
-        match PRINCIPAL_REGEX.captures(string) {
-            Some(captures) => {
-                let principals = captures
-                    .name("principals")
-                    .unwrap()
-                    .as_str()
-                    .replace('\n', "")
-                    .split(',')
-                    .map(|s| s.trim().to_string())
-                    .filter(|s| !s.is_empty())
-                    .collect::<Vec<String>>();
-                Ok(Principals { principals })
-            }
-            None => Err(()),
-        }
-    }
-
-    type Err = (); // TODO: Provide a concrete type for the Err associated type
-}
-
-#[test]
-fn test_parse_principals() {
-    let principals = "ifprincipals { alice, bob }".parse::<Principals>().unwrap();
-    assert_eq!(principals.principals, vec!["alice", "bob"]);
-
-    let principals = "ifprincipals { \n  alice,\n  bob,\n}"
-        .parse::<Principals>()
-        .unwrap();
-    assert_eq!(principals.principals, vec!["alice", "bob"]);
-}
-
-/// Initiate the principals list
-fn initiate_principals(indexer: &Indexer, locator: &Locator) -> Principals {
-    let mut principals = Principals::new_empty();
-    // TODO: Implement logic to extract principals from block comments with comment_ranges.block_comments()
-    for range in indexer.comment_ranges() {
-        let comment = locator.slice(range).replace("#", "");
-        if let Ok(_principals) = comment.parse::<Principals>() {
-            principals.concat(&_principals);
-        }
-    }
-
-    return principals;
-}
-
-#[cfg(test)]
-mod initiate_principals_tests {
-    use super::*;
-
-    #[test]
-    fn test_initiate_principals_with_principals() {
-        use ruff_python_parser::tokenize;
-        use ruff_python_parser::Mode;
-
-        let source: &str = r#"
-# ifprincipals { alice, bob }
-
-# This is a comment
-x = 1
-"#;
-        let tokens = tokenize(source, Mode::Module);
-        let locator = Locator::new(source);
-        let indexer = Indexer::from_tokens(&tokens, &locator);
-        let principals = initiate_principals(&indexer, &locator);
-        assert_eq!(principals, Principals::new_from_str(vec!["alice", "bob"]));
-    }
-
-    #[test]
-    fn test_initiate_principals_with_principals_not_first_comment() {
-        use ruff_python_parser::tokenize;
-        use ruff_python_parser::Mode;
-
-        let source: &str = r#"
-# This is a comment
-# ifprincipals { alice, bob }
-
-x = 1
-"#;
-        let tokens = tokenize(source, Mode::Module);
-        let locator = Locator::new(source);
-        let indexer = Indexer::from_tokens(&tokens, &locator);
-        let principals = initiate_principals(&indexer, &locator);
-        assert_eq!(principals, Principals::new_from_str(vec!["alice", "bob"]));
-    }
-
-    #[test]
-    fn test_initiate_principals_no_principals() {
-        use ruff_python_parser::tokenize;
-        use ruff_python_parser::Mode;
-
-        let source: &str = r#"
-# This is a comment
-x = 1
-"#;
-        let tokens = tokenize(source, Mode::Module);
-        let locator = Locator::new(source);
-        let indexer = Indexer::from_tokens(&tokens, &locator);
-        let principals = initiate_principals(&indexer, &locator);
-        let empty: Vec<String> = vec![];
-        assert_eq!(principals, Principals::new(empty));
-    }
-
-    #[test]
-    fn test_initiate_principals_multiple_principals_concat() {
-        use ruff_python_parser::tokenize;
-        use ruff_python_parser::Mode;
-
-        let source: &str = r#"
-# ifprincipals { alice, bob }
-# ifprincipals { charlie, dean }
-# This is a comment
-x = 1
-"#;
-        let tokens = tokenize(source, Mode::Module);
-        let locator = Locator::new(source);
-        let indexer = Indexer::from_tokens(&tokens, &locator);
-        let principals = initiate_principals(&indexer, &locator);
-        assert_eq!(
-            principals,
-            Principals::new_from_str(vec!["alice", "bob", "charlie", "dean"])
-        );
     }
 }
