+++ crates/ruff_linter/src/checkers/information_flow/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod information_flow_state;
+
+pub(crate) mod label;
+
+pub(crate) mod principals;
